requirements.txt
python-daemon
win10toast
